types: move ResolvedType into resolved.rs
Code move only.
